+++ src/components/MetricAvailability/MetricAvailabilty.types.js
@@ -1,0 +1,3 @@
+// @flow
+
+export type Props = {||};
