--- src/components/MetricFilter/index.js
@@ -1,1 +1,0 @@
-export { MetricFilter } from "./MetricFilter"
