--- src/components/MetricTextSearch/index.js
@@ -1,1 +1,0 @@
-export { MetricTextSearch, searchContent } from "./MetricTextSearch"
